fix: should resolve setup file correctly when setupFile is pure es module
