+++ e2e/setup/fixtures/package-name/test-setup-esm-fixtures/index.js
@@ -1,0 +1,2 @@
+process.env.RETEST_SETUP_FLAG = '1';
+process.env.NODE_ENV = 'rstest:production';
